feat: add PostHog to Braze export
Exports $set properties and events based on the config.exportEvents
comma separated list. If the list is empty, no events will be exported.
I figure that you probably want to be a little more selective about
what you export, so I've made it opt-in. We could add a '*' option if
you really do want to export everything.